Refactor to standalone diff manager
